+++ HonorAmongThieves/Globals/Startup.cs
@@ -1,12 +1,6 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using HonorAmongThieves.Hubs;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.HttpsPolicy;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
@@ -36,7 +30,7 @@ namespace HonorAmongThieves
             services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
 
             services.AddSignalR();
-            services.AddSingleton<Game.Heist.Lobby>();
+            services.AddSingleton<Heist.GameLogic.Lobby>();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
@@ -53,7 +47,7 @@ namespace HonorAmongThieves
 
             app.UseSignalR(routes =>
             {
-                routes.MapHub<HeistHub>("/heistHub");
+                routes.MapHub<Heist.HeistHub>("/heistHub");
             });
 
             app.UseMvc();
